hxat/requirements/base.txt
Django~=3.2.23
channels~=3.0.5
channels_redis==4.1.0
coverage==7.3.2
coverage-badge==1.1.0
django-bootstrap3==23.4
django-braces==1.15.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-log-request-id==2.1.0
djangorestframework==3.14.0
lti==0.9.5
psycopg2-binary>=2.8.4
pyjwt[crypto]==2.8.0
python-dateutil==2.8.2
requests==2.31.0
python_dotenv==1.0.0
WhiteNoise==6.6.0
git+https://github.com/Harvard-ATG/media_management_sdk.git@v0.2.0
